pull: extract bottle publishing code
and add tests
